them option tinh thanh pho, keo duoc data thanh pho

+++ src/main/java/com/dinhpu/m4casestudy/services/IServices.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface IServices<T> {
     public List<T> findAll();
-    public T findById(Integer id);
+    public T findById(Long id);
     public T save(T model);
-    public T remove(Integer id);
+    public T remove(Long id);
 
 }
